+++ AndroidManifest.xml
@@ -44,16 +44,6 @@
         </service>
 
         <activity
-            android:name=".activity.MainActivity"
-            android:label="@string/app_name"
-            android:theme="@android:style/Theme.Black.NoTitleBar" >
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-
-        </activity>
-        <activity
             android:name=".activity.SettingsActivity"
             android:label="@string/app_name"
             android:theme="@style/myAppTheme" />
@@ -68,7 +58,9 @@
         <activity
             android:name=".activity.ListSearchTabActivity"
             android:label="@string/app_name"
-            android:theme="@style/myAppTheme" />
+            android:theme="@style/myAppTheme"
+            android:exported="true"
+            />
 
         <activity
             android:name=".activity.MessagesTabActivity"
