+++ res/layout/starred_hosts_tab.xml
@@ -2,35 +2,37 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical" 
-    android:id="@+id/starredHostsTab"
-     >
+    android:orientation="vertical"
+    android:id="@+id/starredHostsTab">
 
     <include layout="@layout/toolbar" />
+
     <android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:id="@+id/drawer_layout">
-    <!-- activity view -->
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:background="#fff"
-        android:layout_height="match_parent">
+        <!-- activity view -->
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:background="#fff"
+            android:layout_height="match_parent">
 
-    <TextView
-        android:id="@+id/lblNoStarredHosts"
-    	android:layout_width="fill_parent"
-	    android:layout_height="wrap_content"
-	    android:text="@string/no_starred_hosts"
-	    android:visibility="gone"
-        />	
-    <ListView
-    	android:layout_width="fill_parent"
-	    android:layout_height="wrap_content"
-	    android:id="@+id/lstStarredHosts"
-	    android:divider="#FF000000" />
-    </RelativeLayout>
-    <include layout="@layout/navigation_drawer" />
-</android.support.v4.widget.DrawerLayout>/
+            <TextView
+                android:id="@+id/lblNoStarredHosts"
+                android:layout_width="fill_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/no_starred_hosts"
+                android:visibility="gone" />
+
+            <ListView
+                android:layout_width="fill_parent"
+                android:layout_height="wrap_content"
+                android:id="@+id/lstStarredHosts"
+                />
+        </RelativeLayout>
+
+        <include layout="@layout/navigation_drawer" />
+    </android.support.v4.widget.DrawerLayout>
+
 
 </LinearLayout>
