+++ src/fi/bitrite/android/ws/activity/StarredHostTabActivity.java
@@ -17,7 +17,6 @@ import fi.bitrite.android.ws.model.Host;
 import fi.bitrite.android.ws.model.HostBriefInfo;
 import fi.bitrite.android.ws.persistence.StarredHostDao;
 import fi.bitrite.android.ws.persistence.impl.StarredHostDaoImpl;
-
 import java.util.List;
 
 public class StarredHostTabActivity extends WSBaseActivity
@@ -26,7 +25,7 @@ public class StarredHostTabActivity extends WSBaseActivity
     private static final int CONTEXT_MENU_UPDATE = 0;
     private static final int CONTEXT_MENU_DELETE = 1;
 
-    ListView starredHostsList;
+    ListView starredHostList;
     TextView noStarredHostsLabel;
 
     private StarredHostDao starredHostDao = new StarredHostDaoImpl();
@@ -37,10 +36,10 @@ public class StarredHostTabActivity extends WSBaseActivity
         setContentView(R.layout.starred_hosts_tab);
         super.onCreate(savedInstanceState);
 
-        starredHostsList = (ListView)findViewById(R.id.lstStarredHosts);
+        starredHostList = (ListView)findViewById(R.id.lstStarredHosts);
         noStarredHostsLabel = (TextView)findViewById(R.id.lblNoStarredHosts);
 
-        registerForContextMenu(starredHostsList);
+        registerForContextMenu(starredHostList);
     }
 
     private void setupStarredHostsList() {
@@ -48,14 +47,14 @@ public class StarredHostTabActivity extends WSBaseActivity
 
         if (starredHosts.size() == 0) {
             noStarredHostsLabel.setVisibility(View.VISIBLE);
-            starredHostsList.setVisibility(View.GONE);
+            starredHostList.setVisibility(View.GONE);
         } else {
             noStarredHostsLabel.setVisibility(View.GONE);
-            starredHostsList.setVisibility(View.VISIBLE);
+            starredHostList.setVisibility(View.VISIBLE);
 
-            starredHostsList.setAdapter(new HostListAdapter(this, R.layout.host_list_item, starredHosts));
+            starredHostList.setAdapter(new HostListAdapter(this, R.layout.host_list_item, starredHosts));
 
-            starredHostsList.setOnItemClickListener(new OnItemClickListener() {
+            starredHostList.setOnItemClickListener(new OnItemClickListener() {
                 public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                     Intent i = new Intent(StarredHostTabActivity.this, HostInformationActivity.class);
                     HostBriefInfo selectedHost = starredHosts.get(position);
@@ -69,7 +68,7 @@ public class StarredHostTabActivity extends WSBaseActivity
 
     @Override
     public void onCreateContextMenu(ContextMenu menu, View view, ContextMenuInfo menuInfo) {
-        if (view.getId() == starredHostsList.getId()) {
+        if (view.getId() == starredHostList.getId()) {
             AdapterView.AdapterContextMenuInfo info = (AdapterView.AdapterContextMenuInfo) menuInfo;
             menu.setHeaderTitle(starredHosts.get(info.position).getFullname());
             menu.add(Menu.NONE, CONTEXT_MENU_UPDATE, 0, R.string.update);
